feat: v2-v4 — apartados, movimientos, sidebar colapsable y sistema de temas
Cambios acumulados de los updates v2, v3 y v4:

v2 — Apartados y flujo de almacén
- Apartar componentes desde la vista de artículos y almacén
- Vigencia fija de 7 días, proyecto obligatorio para apartar
- Stock visible refleja total − apartado reservado
- Corrección de referencia obsoleta a proveedor en API de entradas

v3 — Movimiento de componentes, reportes y ubicaciones en artículos
- API atómica POST /api/movimientos con transacción Prisma
- Modal MoverComponenteModal con cascada de dropdowns
- Columna Ubicación(es) en catálogo de artículos (UbicacionesBadge)
- Historial de movimientos en detalle de artículo
- Reporte de movimientos con filtros, paginación y exportación Excel
- Modelo Movimiento en schema.prisma con índices y relaciones nombradas
- Notificaciones in-app a propietarios de apartados afectados

v4 — Sistema de temas Light/Dark + tipografía Barlow Semi Condensed
- Barlow Semi Condensed via next/font (var(--font-barlow))
- Variables CSS completas para tema light (púrpura/carmesí) y dark (cyan)
- ThemeProvider con persistencia en localStorage y fallback a preferencia del SO
- ThemeToggle pill 64×28px (Sun/Moon) en Topbar y login
- Script anti-FOUC inline en <head> para evitar parpadeo al cargar
- Sidebar colapsable (desktop 64px / 240px) con localStorage y transición 200ms
- Colores hardcodeados reemplazados por CSS variables y color-mix() en toda la app
- Badges temáticos via clases CSS badge-success/warning/danger/info/purple/cyan
- Login con gradiente de marca en light, glassmorphism en dark
- Hook useChartColors para futuros componentes Recharts

+++ app/api/dashboard/route.ts
@@ -45,11 +45,11 @@ export async function GET() {
     }),
     prisma.proyecto.count({ where: { estado: 'ACTIVO' } }),
     prisma.$queryRaw`
-      SELECT DATE("fecha") as dia, COUNT(*) as total, 'ENTRADA' as tipo
+      SELECT DATE("fecha") as dia, COUNT(*)::int as total, 'ENTRADA' as tipo
       FROM "Entrada" WHERE "fecha" >= ${subDays(ahora, 30)}
       GROUP BY DATE("fecha")
       UNION ALL
-      SELECT DATE("fecha") as dia, COUNT(*) as total, 'SALIDA' as tipo
+      SELECT DATE("fecha") as dia, COUNT(*)::int as total, 'SALIDA' as tipo
       FROM "Salida" WHERE "fecha" >= ${subDays(ahora, 30)}
       GROUP BY DATE("fecha")
       ORDER BY dia ASC
